fix segfault
